new js workflow items - eslint, prettier, husky, lint-staged - remove lib folder - move test files to test folder
