+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules/
 coverage/
 .idea/
+lib/
+npm-debug.log
