--- lib/actions.js
@@ -1,50 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.createAction = createAction;
-
-
-// App-level request
-var DATA_REQUESTED = 'application/DATA_REQUESTED';
-var PERIODIC_DATA_REQUESTED = 'application/PERIODIC_DATA_REQUESTED';
-var PERIODIC_TERMINATION_REQUESTED = 'application/PERIODIC_TERMINATION_REQUESTED';
-var DATA_REQUESTED_USE_LATEST = 'application/DATA_REQUESTED_USE_LATEST';
-
-// System responses
-var PERIODIC_TERMINATION_SUCCEEDED = 'sagas/PERIODIC_TERMINATION_SUCCEEDED';
-
-// System requests for fetching/updating
-var FETCH_REQUESTED = 'net/FETCH_REQUESTED';
-var TRANSIENT_FETCH_REQUESTED = 'net/TRANSIENT_FETCH_REQUESTED';
-var KEY_REMOVAL_REQUESTED = 'net/KEY_REMOVAL_REQUESTED';
-
-// System responses to fetch requests
-var FETCH_RESULT_RECEIVED = 'net/FETCH_RESULT_RECEIVED';
-var TRANSIENT_FETCH_RESULT_RECEIVED = 'net/TRANSIENT_FETCH_RESULT_RECEIVED';
-var FETCH_FAILED = 'net/FETCH_FAILED';
-var FETCH_TIMED_OUT = 'net/FETCH_TIMED_OUT';
-var TRANSIENT_FETCH_FAILED = 'net/TRANSIENT_FETCH_FAILED';
-var FETCH_TRY_FAILED = 'net/FETCH_TRY_FAILED';
-
-function createAction(type, payload) {
-	return Object.assign({}, { type: type }, payload);
-}
-
-exports.default = {
-	DATA_REQUESTED: DATA_REQUESTED,
-	PERIODIC_DATA_REQUESTED: PERIODIC_DATA_REQUESTED,
-	PERIODIC_TERMINATION_REQUESTED: PERIODIC_TERMINATION_REQUESTED,
-	DATA_REQUESTED_USE_LATEST: DATA_REQUESTED_USE_LATEST,
-	PERIODIC_TERMINATION_SUCCEEDED: PERIODIC_TERMINATION_SUCCEEDED,
-	FETCH_REQUESTED: FETCH_REQUESTED,
-	TRANSIENT_FETCH_REQUESTED: TRANSIENT_FETCH_REQUESTED,
-	KEY_REMOVAL_REQUESTED: KEY_REMOVAL_REQUESTED,
-	FETCH_FAILED: FETCH_FAILED,
-	FETCH_TIMED_OUT: FETCH_TIMED_OUT,
-	TRANSIENT_FETCH_FAILED: TRANSIENT_FETCH_FAILED,
-	FETCH_TRY_FAILED: FETCH_TRY_FAILED,
-	FETCH_RESULT_RECEIVED: FETCH_RESULT_RECEIVED,
-	TRANSIENT_FETCH_RESULT_RECEIVED: TRANSIENT_FETCH_RESULT_RECEIVED
-};
