--- lib/fetchSaga.js
@@ -1,484 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-
-var _regenerator = require('babel-runtime/regenerator');
-
-var _regenerator2 = _interopRequireDefault(_regenerator);
-
-exports.default = fetchSaga;
-
-var _reduxSaga = require('redux-saga');
-
-var _effects = require('redux-saga/effects');
-
-var _lodash = require('lodash');
-
-var _lodash2 = _interopRequireDefault(_lodash);
-
-var _fetchService = require('./services/fetchService');
-
-var _actions = require('./actions');
-
-var _actions2 = _interopRequireDefault(_actions);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-var _marked = [fetchData, fetchOnce, fetchDataLoop, fetchDataRecurring, fetchSaga].map(_regenerator2.default.mark);
-
-/**
- * Oauth token as generated by OWIN.NET.
- * 
- */
-
-
-/**
- * modelName - The key that is used to locate the request config in apis.js and also to place the result in the redux store
- * headers - (optional) An object as key/value pairs of headers to be sent with the request
- * queryParams - (optional) An object as key/value pairs to be added to query as query params
- * noStore - (optional) If true, make the request but do not store in redux. Can be used with take & friends for side effects
- * period - (optional) How often to re-fetch when used in a recurring fetch scenario
- * taskId - (optional) A pre-generated (by your application) id to be used to cancel a recurring task at a later time
- * 
- */
-var logger = void 0;
-var models = void 0;
-var tokenAccessFunction = void 0;
-var errorFunction = void 0;
-
-/**
- * Construct a request based on the provided action, make a request with a configurable retry,
- * and handle errors, logging and dispatching all steps.
- * 
- * @param {FetchAction} action - An action with the request configuration
- */
-function fetchData(action) {
-	var baseConfig, headers, fetchConfig, store, tryLimit, tryCount, didFail, didTimeOut, lastError, oauthToken, _ref, fetchResult, timedOut, errorObject, _errorObject, _errorObject2, _fetchResult;
-
-	return _regenerator2.default.wrap(function fetchData$(_context) {
-		while (1) {
-			switch (_context.prev = _context.next) {
-				case 0:
-					if (!(!action || !action.modelName)) {
-						_context.next = 2;
-						break;
-					}
-
-					throw new Error("'modelName' config parameter is required for fetchData");
-
-				case 2:
-
-					// Get fetch parameters from global fetch dictionary using the modelName passed in to locate them
-					// Combine parameters from global dictionary with any passed in - locals override dictionary
-					baseConfig = _lodash2.default.get(models, action.modelName);
-
-					if (baseConfig) {
-						_context.next = 5;
-						break;
-					}
-
-					throw new Error('Cannot find \'' + action.modelName + '\' model in model dictionary');
-
-				case 5:
-					headers = Object.assign({}, baseConfig.headers, action.headers);
-					fetchConfig = Object.assign({}, baseConfig, {
-						headers: headers
-					});
-
-					if (action.body || baseConfig.body) {
-						// If the body is a string, we are assuming it's an application/x-www-form-urlencoded
-						if (typeof action.body === 'string') {
-							fetchConfig.body = action.body;
-						} else {
-							fetchConfig.body = Object.assign({}, baseConfig.body, action.body);
-						}
-					}
-					fetchConfig.queryParams = Object.assign({}, baseConfig.queryParams, action.queryParams);
-
-					// substitute parameterized query path references with values from store
-					// TODO: validate the path exists in the store
-
-					if (!/{{.+}}/.test(fetchConfig.path)) {
-						_context.next = 14;
-						break;
-					}
-
-					_context.next = 12;
-					return (0, _effects.select)(function (state) {
-						return state;
-					});
-
-				case 12:
-					store = _context.sent;
-
-					fetchConfig.path = fetchConfig.path.replace(/{{(.+?)}}/, function (matches, backref) {
-						return _lodash2.default.get(store, backref);
-					});
-
-				case 14:
-
-					// Configure retry
-					tryLimit = action.noRetry ? 0 : 4;
-					tryCount = 0;
-					didFail = void 0;
-					didTimeOut = void 0;
-					lastError = '';
-
-					// Run retry loop
-
-				case 19:
-					didFail = false;
-					didTimeOut = false;
-
-					tryCount++;
-					// Indicate fetch action has begun
-					_context.next = 24;
-					return (0, _effects.put)((0, _actions.createAction)(action.noStore ? _actions2.default.TRANSIENT_FETCH_REQUESTED : _actions2.default.FETCH_REQUESTED, {
-						modelName: action.modelName
-					}));
-
-				case 24:
-					_context.prev = 24;
-					_context.next = 27;
-					return (0, _effects.call)(tokenAccessFunction, action.modelName);
-
-				case 27:
-					oauthToken = _context.sent;
-
-					if (oauthToken && oauthToken.access_token) {
-						fetchConfig.headers['Authorization'] = 'Bearer ' + oauthToken.access_token;
-					}
-					_context.next = 31;
-					return (0, _effects.race)({
-						fetchResult: (0, _effects.call)(_fetchService.doFetch, fetchConfig),
-						timedOut: (0, _effects.call)(_reduxSaga.delay, action.timeLimit ? action.timeLimit : 30000)
-					});
-
-				case 31:
-					_ref = _context.sent;
-					fetchResult = _ref.fetchResult;
-					timedOut = _ref.timedOut;
-
-					if (!(fetchResult && !(fetchResult.title && fetchResult.title === 'Error'))) {
-						_context.next = 39;
-						break;
-					}
-
-					_context.next = 37;
-					return (0, _effects.put)((0, _actions.createAction)(action.noStore ? _actions2.default.TRANSIENT_FETCH_RESULT_RECEIVED : _actions2.default.FETCH_RESULT_RECEIVED, {
-						data: fetchResult,
-						modelName: action.modelName
-					}));
-
-				case 37:
-					_context.next = 51;
-					break;
-
-				case 39:
-					if (!timedOut) {
-						_context.next = 47;
-						break;
-					}
-
-					_context.next = 42;
-					return (0, _effects.put)((0, _actions.createAction)(_actions2.default.FETCH_TIMED_OUT, {
-						modelName: action.modelName
-					}));
-
-				case 42:
-					didTimeOut = true;
-					errorObject = {
-						type: _actions2.default.FETCH_TIMED_OUT,
-						modelName: action.modelName,
-						errorData: fetchResult
-					};
-					throw new Error(JSON.stringify(errorObject));
-
-				case 47:
-					_context.next = 49;
-					return (0, _effects.put)((0, _actions.createAction)(_actions2.default.FETCH_TRY_FAILED, {
-						modelName: action.modelName,
-						errorData: fetchResult
-					}));
-
-				case 49:
-					_errorObject = {
-						type: _actions2.default.FETCH_TRY_FAILED,
-						modelName: action.modelName,
-						errorData: fetchResult
-					};
-					throw new Error(JSON.stringify(_errorObject));
-
-				case 51:
-					_context.next = 64;
-					break;
-
-				case 53:
-					_context.prev = 53;
-					_context.t0 = _context['catch'](24);
-					_errorObject2 = JSON.parse(_context.t0.message);
-					_fetchResult = _errorObject2.errorData;
-
-					// Don't do anything with 401 errors
-					// And some errors don't have fetch results associated with them
-
-					if (errorFunction && (_fetchResult ? _fetchResult.code != 401 : true)) {
-						errorFunction(_context.t0.message);
-					}
-					didFail = true;
-					lastError = _context.t0;
-					logger('fetchData fail');
-					logger(_context.t0);
-					_context.next = 64;
-					return (0, _effects.call)(_reduxSaga.delay, 2 ^ tryCount * 100);
-
-				case 64:
-					if (tryCount < tryLimit && didFail) {
-						_context.next = 19;
-						break;
-					}
-
-				case 65:
-					if (!(tryCount === tryLimit && didFail)) {
-						_context.next = 71;
-						break;
-					}
-
-					if (didTimeOut) {
-						_context.next = 69;
-						break;
-					}
-
-					_context.next = 69;
-					return (0, _effects.put)((0, _actions.createAction)(_actions2.default.FETCH_FAILED, { modelName: action.modelName }));
-
-				case 69:
-					logger('fetchData retry fail');
-					logger(lastError);
-
-				case 71:
-				case 'end':
-					return _context.stop();
-			}
-		}
-	}, _marked[0], this, [[24, 53]]);
-}
-
-/**
- * Call the fetchData saga exactly one time (keeping in mind fetchData has retries by default)
- * 
- * @param {FetchAction} action - An action with the request configuration
- */
-function fetchOnce(action) {
-	return _regenerator2.default.wrap(function fetchOnce$(_context2) {
-		while (1) {
-			switch (_context2.prev = _context2.next) {
-				case 0:
-					_context2.next = 2;
-					return (0, _effects.call)(fetchData, action);
-
-				case 2:
-				case 'end':
-					return _context2.stop();
-			}
-		}
-	}, _marked[1], this);
-}
-
-/**
- * The loop saga that makes the request every {config.period} milliseconds until
- * cancelled
- * 
- * @param {FetchAction} action - An action with the request configuration
- */
-function fetchDataLoop(action) {
-	return _regenerator2.default.wrap(function fetchDataLoop$(_context3) {
-		while (1) {
-			switch (_context3.prev = _context3.next) {
-				case 0:
-					_context3.prev = 0;
-
-				case 1:
-					if (!true) {
-						_context3.next = 8;
-						break;
-					}
-
-					_context3.next = 4;
-					return (0, _effects.call)(fetchData, action);
-
-				case 4:
-					_context3.next = 6;
-					return (0, _effects.call)(_reduxSaga.delay, action.period);
-
-				case 6:
-					_context3.next = 1;
-					break;
-
-				case 8:
-					_context3.prev = 8;
-					_context3.next = 11;
-					return (0, _effects.cancelled)();
-
-				case 11:
-					if (!_context3.sent) {
-						_context3.next = 14;
-						break;
-					}
-
-					_context3.next = 14;
-					return (0, _effects.put)((0, _actions.createAction)(_actions2.default.PERIODIC_TERMINATION_SUCCEEDED, { modelName: action.modelName }));
-
-				case 14:
-					return _context3.finish(8);
-
-				case 15:
-				case 'end':
-					return _context3.stop();
-			}
-		}
-	}, _marked[2], this, [[0,, 8, 15]]);
-}
-
-function matchesTerminationAction(incomingAction, fetchAction) {
-	return incomingAction.type === _actions2.default.PERIODIC_TERMINATION_REQUESTED && incomingAction.taskId === fetchAction.taskId;
-}
-
-/**
- * Call the fetchData saga every {config.period} milliseconds. This saga requires the 'period' and 'taskId' properties
- * on the action parameter.
- * 
- * @param {FetchAction} action - An action with the request configuration 
- */
-function fetchDataRecurring(action) {
-	var bgSyncTask;
-	return _regenerator2.default.wrap(function fetchDataRecurring$(_context4) {
-		while (1) {
-			switch (_context4.prev = _context4.next) {
-				case 0:
-					if (!(!action || !action.period)) {
-						_context4.next = 2;
-						break;
-					}
-
-					throw new Error("'period' config parameter is required for fetchDataRecurring");
-
-				case 2:
-					if (!(!action || !action.taskId)) {
-						_context4.next = 4;
-						break;
-					}
-
-					throw new Error("'taskId' config parameter is required for fetchDataRecurring");
-
-				case 4:
-					_context4.next = 6;
-					return (0, _effects.fork)(fetchDataLoop, action);
-
-				case 6:
-					bgSyncTask = _context4.sent;
-					_context4.next = 9;
-					return (0, _effects.take)(function (incomingAction) {
-						return matchesTerminationAction(incomingAction, action);
-					});
-
-				case 9:
-					_context4.next = 11;
-					return (0, _effects.cancel)(bgSyncTask);
-
-				case 11:
-				case 'end':
-					return _context4.stop();
-			}
-		}
-	}, _marked[3], this);
-}
-
-/**
- * A default logger function that logs to the console. Used if no other logger is provided
- * 
- * @param {string} message - The message to log
- */
-var consoleLogger = function consoleLogger(message) {
-	console.debug(message);
-};
-
-var tokenAccess = function tokenAccess() {
-	return undefined;
-};
-
-/**
- * The main saga for fetching data. Must be initialized with an object representing the models that can be fetched
- * and an API root to prepend to any partial URLs specified in the models object. A logger should normally be provided
- * as well.
- * 
- * Models object require a form as follows (with optional nested models):
- * {
- * 	fryModel: {
- * 		path: '/api/Foo'
- * 	},
- * 	groupOfModels: {
- * 		leelaModel: {
- * 			path: '/api/Bar'
- * 		},
- * 		benderModel: {
- * 			path: '/api/Baz'
- * 		}
- * 	}
- * }
- * 
- * Models are referenced in the actions.DATA_REQUESTED action by path, i.e.
- * { type: actions.DATA_REQUESTED, { modelName: 'fryModel' } }
- * -- or --
- * { type: actions.DATA_REQUESTED, { modelName: 'groupOfModels.leelaModel' } }
- * 
- * @export
- * @param {Object} modelsParam - An object indicating the APIs available in a application with which to make requests
- * @param {string} apiRootParam - A url to which partial URLs are appended (i.e.) 'https://myapp.com'
- * @param {TokenAccessFunction} [tokenAccessParam=tokenAccess] - function that returns an optional OAuth token
- * @param {ErrorFunction} errorParam  - A function to perform on errors
- * @param {LoggerFunction} [loggerParam=consoleLogger] - A function that accepts a string and logs it real good
- */
-function fetchSaga(modelsParam, apiRootParam) {
-	var tokenAccessParam = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : tokenAccess;
-	var errorParam = arguments[3];
-	var loggerParam = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : consoleLogger;
-	return _regenerator2.default.wrap(function fetchSaga$(_context5) {
-		while (1) {
-			switch (_context5.prev = _context5.next) {
-				case 0:
-					if (modelsParam) {
-						_context5.next = 2;
-						break;
-					}
-
-					throw new Error("'modelsParam' is required for fetchSaga");
-
-				case 2:
-					(0, _fetchService.setApiRoot)(apiRootParam);
-					logger = loggerParam;
-					logger('logger set to ' + logger.name);
-					models = modelsParam;
-					errorFunction = errorParam;
-					tokenAccessFunction = tokenAccessParam;
-
-					_context5.next = 10;
-					return (0, _effects.takeEvery)(_actions2.default.DATA_REQUESTED, fetchOnce);
-
-				case 10:
-					_context5.next = 12;
-					return (0, _effects.takeEvery)(_actions2.default.PERIODIC_DATA_REQUESTED, fetchDataRecurring);
-
-				case 12:
-					_context5.next = 14;
-					return (0, _effects.takeLatest)(_actions2.default.DATA_REQUESTED_USE_LATEST, fetchOnce);
-
-				case 14:
-				case 'end':
-					return _context5.stop();
-			}
-		}
-	}, _marked[4], this);
-}
