--- lib/index.js
@@ -1,27 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.sagas = exports.reducers = exports.actions = undefined;
-
-var _actions = require('./actions');
-
-var _actions2 = _interopRequireDefault(_actions);
-
-var _fetchReducer = require('./fetchReducer');
-
-var _fetchReducer2 = _interopRequireDefault(_fetchReducer);
-
-var _fetchSaga = require('./fetchSaga');
-
-var _fetchSaga2 = _interopRequireDefault(_fetchSaga);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-var reducers = { fetchReducer: _fetchReducer2.default };
-
-var sagas = { fetchSaga: _fetchSaga2.default };
-exports.actions = _actions2.default;
-exports.reducers = reducers;
-exports.sagas = sagas;
