+++ src/actions.js
@@ -3,8 +3,7 @@
 // App-level request
 const DATA_REQUESTED = 'application/DATA_REQUESTED'
 const PERIODIC_DATA_REQUESTED = 'application/PERIODIC_DATA_REQUESTED'
-const PERIODIC_TERMINATION_REQUESTED =
-	'application/PERIODIC_TERMINATION_REQUESTED'
+const PERIODIC_TERMINATION_REQUESTED = 'application/PERIODIC_TERMINATION_REQUESTED'
 const DATA_REQUESTED_USE_LATEST = 'application/DATA_REQUESTED_USE_LATEST'
 
 // System responses
